Few changes: - Move all native queries into a StatisticsRepository (because all other queries in the app use the DBAL, so if we change the kind of db all native queries will be at the same place) - Optimized few queries
