+++ symfony/src/Repository/MessageRepository.php
@@ -16,8 +16,6 @@ use Doctrine\Common\Persistence\Mapping\MappingException;
 use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
-use Doctrine\ORM\Query\ResultSetMapping;
-use Doctrine\ORM\QueryBuilder;
 
 class MessageRepository extends BaseRepository
 {
@@ -212,84 +210,6 @@ class MessageRepository extends BaseRepository
             ->getSingleScalarResult();
 
         return !$result;
-    }
-
-    /**
-     * Return all sent messages and group by kind (communication.type)
-     *
-     * @param \DateTime $from
-     * @param \DateTime $to
-     * @return array
-     */
-    public function getNumberOfSentMessagesByKind(\DateTime $from, \DateTime $to)
-    {
-        $sql = 'select c.type, count(*)
-                from message m join communication c
-                on m.communication_id = c.id
-                where c.created_at > :fromDate
-                and c.created_at < :toDate
-                group by c.type;';
-
-        $rsm = new ResultSetMapping();
-        $rsm->addScalarResult('type', 'type')
-            ->addScalarResult('count(*)', 'count', 'integer');
-
-        return $this->_em
-            ->createNativeQuery($sql, $rsm)
-            ->setParameter('fromDate', $from)
-            ->setParameter('toDate', $to)
-            ->getResult();
-    }
-
-    /**
-     * Return all triggered volounteers
-     * @param \DateTime $from
-     * @param \DateTime $to
-     * @return array
-     * @throws NonUniqueResultException
-     * @throws \Doctrine\ORM\NoResultException
-     */
-    public function getNumberOfTriggeredVolounteers(\DateTime $from, \DateTime $to)
-    {
-        $sql = 'select count(distinct m.volunteer_id) volounteers
-                from message m
-                join communication c on m.communication_id = c.id
-                where c.created_at > :fromDate
-                and c.created_at <= :toDate;';
-
-        $rsm = new ResultSetMapping();
-        $rsm->addScalarResult('volounteers', 'volounteers', 'integer');
-
-        return $this->_em
-            ->createNativeQuery($sql, $rsm)
-            ->setParameter('fromDate', $from)
-            ->setParameter('toDate', $to)
-            ->getSingleResult();
-    }
-
-    /**
-     * Return all answers received
-     *
-     * @return QueryBuilder
-     * @throws NonUniqueResultException
-     * @throws \Doctrine\ORM\NoResultException
-     */
-    public function getNumberOfAnswersReceived(\DateTime $from, \DateTime $to)
-    {
-        $sql = 'select count(*) answers from message m
-                join communication c on m.communication_id = c.id
-                join answer a on a.message_id = m.id
-                where c.created_at > :fromDate
-                and c.created_at <= :toDate';
-
-        $rsm = new ResultSetMapping();
-        $rsm->addScalarResult('answers', 'answers', 'integer');
-
-        return $this->_em
-            ->createNativeQuery($sql, $rsm)
-            ->setParameter(':fromDate', $from)
-            ->setParameter(':toDate', $to)
-            ->getSingleResult();
     }
 
     /**
